PSY Methodology Crisis Identification for Crises Periods, With a Focus on Market Indices and Yield Spread
Observing the effects of crises periods for the Japanese and the US economy, observing how the effect of the crisis effects the market index performances for the S&P500 and the NASDAQ in the US and for the Nikkei 225 for Japan, whilst also comparing yield spreads for the two economies. Fundamental analysis was also performed for a more complete understanding.